Product Arrivals functionality
This can be used when you order stock and want to allow customers to order the products while you are waiting for the stock to arrive.
You can set when the stock is arriving and it will create a StockStatus showing the date (range) of when the product will be back in stock,
and also allow backordering for this product variant, if it wasn't already allowed.